added detection of union types on user aliases

+++ examples/basic_user_alias.py
@@ -1,4 +1,4 @@
-Number = int | float
+Number = float | int
 
 
 # TODO: allow the user to change the input type on the frontend
